+++ car.php
@@ -2,11 +2,6 @@
 session_start();
 require_once 'config.php';
 
-// Activer l'affichage des erreurs
-ini_set('display_errors', 1);
-ini_set('display_startup_errors', 1);
-error_reporting(E_ALL);
-
 // Récupération de la recherche de l'utilisateur
 $search = strtolower($_GET['search'] ?? '');